pyproject.toml
[tool.poetry]
name = "cw5-vacancies-parsing-to-db"
version = "0.1.0"
description = ""
authors = ["Yerzhan Amirkhanov <yerzhanamirkhanov@gmail.com>"]
readme = "README.md"
packages = [{include = "cw5_vacancies_parsing_to_db"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2 = "^2.9.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
